% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{plot.perm_test}
\alias{plot.perm_test}
\title{Plot method for permutation test objects}
\usage{
\method{plot}{perm_test}(x, bins = 30, col = "lightgray", line_col = "red", ...)
}
\arguments{
\item{x}{An object of class \code{"perm_test"}.}

\item{bins}{Number of histogram bins (default 30).}

\item{col}{Color for histogram bars (default "lightgray").}

\item{line_col}{Color for observed statistic line (default "red").}

\item{...}{Additional arguments passed to \code{hist()}.}
}
\description{
Produces a histogram of the permutation distribution with a vertical line
indicating the observed statistic.
}
